tweaks to Teleop

+++ Misc_Template.h
@@ -75,12 +75,12 @@ float dontMore(float num, float max){
 	return num;
 }
 // return [float] number raised to a non-negative integer power
-float ipow(float base, int power) {
-	float val = 1;
-	
+long int ipow(float base, int power) {
+	long int val = 1;
+
 	for (int i = 0; i < power; i++)
 		val *= base;
-	
+
 	return val;
 }
 #endif // _MISC_TEMPLATE_H_
